main/labs/lab3/solution/lab3/pyproject.toml
[tool.poetry]
name = "lab3"
version = "0.3.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24"
joblib = "^1.3"
scikit-learn = "1.3.2"
requests = "^2.31"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
isort = "^5.12"
black = "^23.11"
pre-commit = "^3.5"
httpx = "^0.25"
ruff = "^0.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
